requirements.txt
Flask==2.3.3
psutil==5.9.5
requests==2.31.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
